# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tuk Tracker API
  version: 1.0.0
  description: Coursework API for tracking tuk-tuks, routes, schedules, and live locations.
servers:
  - url: http://localhost:5000
paths:
...
  /api/routes:
    get:
      summary: List all routes
      responses:
        '200':
          description: Array of route objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
    post:
      summary: Create a route (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteInput'
            examples:
              example:
                value: { "routeCode": "R1", "name": "Downtown Loop", "origin": "Station A", "destination": "Station B", "stops": ["Stop 1","Stop 2"], "distanceKm": 12.5 }
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '401':
          description: Unauthorized
  /api/routes/{id}:
    get:
      summary: Get a route by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Route resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Route not found
...
  /api/history:
    get:
      summary: Query movement history
      parameters:
        - in: query
          name: tukId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Location history points
  /api/tuks/{id}/history:
    get:
      summary: Get movement history for one tuk-tuk
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Tuk movement history
components:
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Route:
      type: object
      properties:
        _id:
          type: string
        routeCode:
          type: string
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        stops:
          type: array
          items:
            type: string
        distanceKm:
          type: number
    RouteInput:
      type: object
      properties:
        routeCode:
          type: string
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        stops:
          type: array
          items:
            type: string
        distanceKm:
          type: number
      required: [routeCode]
...